requirements.txt
dnspython==2.2.1
pp-ez==0.2.0
pymongo==4.3.3
python-dotenv==0.21.0
